moved function to entityutils
